move httpRequest wrapper to static folder
